+++ NEWS.md
@@ -1,4 +1,7 @@
-# QCkit 0.1.0.5 "Work in Progress"
+# QCkit 0.1.1 "Automated Shuttle Pilot"
+2023-11-20
+* Added the function `create_datastore_scipt()`, which given a username and repo for GitHub will generate a draft Script Reference on DataStore based on the information found in the latest Release on GitHub.
+
 24 April 2023
 * fixed bug in `get_custom_flags()`.